+++ lib/gamefic/proxy/plot_pick.rb
@@ -4,7 +4,7 @@ module Gamefic
   class Proxy
     class PlotPick < Base
       def select narrative
-        narrative.plot.pick *args
+        raise? ? narrative.plot.pick!(*args) : narrative.plot.pick(*args)
       end
     end
   end